Merged PR 292: Initial Development
Working implementation of BOS Auth Client Library that provides full set of functionality that the API offers.
Abstracts set up to a single line in Startup.cs through a IServiceCollection extension method.
Includes a read me with the specific codes used for AuthErrors when it is used.

Related work items: #652